+++ example/region/variables.tf
@@ -1,7 +1,9 @@
-variable "domain" {
-  type = string
-}
-
-variable "parameters" {
-  type = map(string)
-}
+variable "domain" { type = string }
+variable "slack_signing_secret" { type = string }
+variable "slack_client_id" { type = string }
+variable "slack_client_secret" { type = string }
+variable "slack_scope" { type = string }
+variable "slack_user_scope" { type = string }
+variable "slack_error_uri" { type = string }
+variable "slack_success_uri" { type = string }
+variable "slack_token" { type = string }
