+++ example/terraform.tf
@@ -28,30 +28,58 @@ provider "aws" {
 #   VARIABLES   #
 #################
 
-variable "domain" { type = string }
-variable "parameters" { type = map(string) }
+variable "domain" {
+  type = string
+}
+
+variable "slack_signing_secret" {
+  type      = string
+  sensitive = true
+}
+
+variable "slack_client_id" {
+  type = string
+}
+
+variable "slack_client_secret" {
+  type      = string
+  sensitive = true
+}
+
+variable "slack_scope" {
+  type = string
+}
+
+variable "slack_user_scope" {
+  type = string
+}
+
+variable "slack_error_uri" {
+  type = string
+}
+
+variable "slack_success_uri" {
+  type = string
+}
+
+variable "slack_token" {
+  type = string
+}
 
 ###############
 #   REGIONS   #
 ###############
 
-# module "eu-west-2" {
-#   providers = { aws = aws.eu-west-2 }
-#   source    = "./region"
-#   domain    = var.domain
-#   secret    = var.secret
-# }
-
 module "us-east-1" {
-  providers  = { aws = aws.us-east-1 }
-  source     = "./region"
-  domain     = var.domain
-  parameters = var.parameters
+  providers            = { aws = aws.us-east-1 }
+  source               = "./region"
+  domain               = var.domain
+  slack_signing_secret = var.slack_signing_secret
+  slack_client_id      = var.slack_client_id
+  slack_client_secret  = var.slack_client_secret
+  slack_scope          = var.slack_scope
+  slack_user_scope     = var.slack_user_scope
+  slack_error_uri      = var.slack_error_uri
+  slack_success_uri    = var.slack_success_uri
+  slack_token          = var.slack_token
 }
-
-# module "us-west-2" {
-#   providers = { aws = aws.us-west-2 }
-#   source    = "./region"
-#   domain    = var.domain
-#   secret    = var.secret
-# }
